Se realizaron los siguientes cambios
se modificó barra de busqueda y se agregó a todas las paginas
se agregó colores a tareas pendientes(warning)/vigentes(success)/eliminadas(danger)
se agregó funcion validar fecha para editar tarea
se modificó historial de tareas, botones restaurar/eliminar
se agregó contador de dias restantes para vencimiento de tareas (tareas pendientes)

+++ app/views/Tareas/historial.php
@@ -3,9 +3,12 @@
 
 <script src="https://kit.fontawesome.com/3ca9a6174a.js" crossorigin="anonymous"></script>
 
-<div class="container mt-4">
 
+<div class="container mt-4 text-center">
     <h2>Historial de Tareas</h2>
+    <br>
+    <?php include __DIR__ . '/../layouts/buscar.php'; ?>
+    <br>
     <table class="table table-striped  text-center">
         <thead>
             <tr>
@@ -21,16 +24,18 @@
         <tbody>
             <?php foreach ($tareas as $tarea): ?>
 
-                <tr class="<?php echo $tarea['tarea_eliminada'] ? 'text-decoration-line-through' : ''; ?>">
+                <tr class="<?php echo $tarea['tarea_eliminada'] ? 'text-decoration-line-through table-danger ' : ''; ?>">
                     <th scope="row"><?php echo htmlspecialchars($tarea['tareas_id']); ?></th>
                     <td><?php echo htmlspecialchars($tarea['tareas_titulo']); ?></td>
                     <td><?php echo htmlspecialchars($tarea['tareas_descripcion']); ?></td>
                     <td><?php echo htmlspecialchars($tarea['tarea_vencimiento']); ?></td>
                     <td><?php echo $tarea['tarea_completada'] ? 'Sí' : 'No'; ?></td>
                     <td><?php echo $tarea['tarea_eliminada'] ? 'Sí' : 'No'; ?></td>
-                    <td>                    
-                    <a href="/TODOapp/public/index.php?accion=eliminar&id=<?php echo $tarea['tareas_id']; ?>" class="btn btn-danger btn-small" data-bs-toggle="popover" data-bs-trigger="hover focus" data-bs-content="Eliminar definitivamente"><i class="fa-solid fa-trash"></i>
-                    <a href="/TODOapp/public/index.php?accion=eliminarLogico&id=<?php echo $tarea['tareas_id']; ?>" class="btn btn-success btn-small" data-bs-toggle="popover" data-bs-trigger="hover focus" data-bs-content="Restaurar la tarea"><i class="fa-solid fa-trash-arrow-up"></i>
+                    <td>
+                        <a href="/TODOapp/public/index.php?accion=eliminar&id=<?php echo $tarea['tareas_id']; ?>" class="btn btn-danger btn-small m-1" data-bs-toggle="popover" data-bs-trigger="hover focus" data-bs-content="Eliminar definitivamente"><i class="fa-solid fa-trash"></i>
+                            <?php if ($tarea['tarea_eliminada']): ?>
+                                <a href="/TODOapp/public/index.php?accion=eliminarLogico&id=<?php echo $tarea['tareas_id']; ?>" class="btn btn-success btn-small" data-bs-toggle="popover" data-bs-trigger="hover focus" data-bs-content="Restaurar la tarea"><i class="fa-solid fa-trash-arrow-up"></i>
+                                <?php endif; ?>
                     </td>
                 </tr>
 
@@ -43,8 +48,8 @@
 <link href="https://getbootstrap.com/docs/5.3/assets/css/docs.css" rel="stylesheet">
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"></script>
 <script>
-const popoverTriggerList = document.querySelectorAll('[data-bs-toggle="popover"]')
-const popoverList = [...popoverTriggerList].map(popoverTriggerEl => new bootstrap.Popover(popoverTriggerEl))
+    const popoverTriggerList = document.querySelectorAll('[data-bs-toggle="popover"]')
+    const popoverList = [...popoverTriggerList].map(popoverTriggerEl => new bootstrap.Popover(popoverTriggerEl))
 </script>
 
 <?php include __DIR__ . '/../layouts/footer.php'; ?>
